refactor(core): lift font design line-metrics to core (shared by docx/pptx/xlsx)
The OS/2 win / hhea design line-metrics table + fontWinLineHeightRatio /
intendedSingleLinePx / correctLineMetrics lived in docx (`font-metrics.ts`), but
the faces it corrects (Meiryo/Meiryo UI, Sakkal Majalla, Times New Roman, Arial)
render identically in a pptx or xlsx text box — so all three formats need the
SAME design line height to avoid under-/over-measuring a substituted face.

Move the module (+ its test) to `core/src/text/line-metrics.ts` and export it
from core. docx keeps a thin re-export at `font-metrics.ts` so its import sites
are unchanged. Pure data + functions, no behavior change (docx + core: 819 tests
green, tsc --build clean). pptx/xlsx wiring follows in separate commits.
